+++ src/view/show_mode/editor/show_ui_widgets/__init__.py
@@ -3,7 +3,7 @@ from model.filter import FilterTypeEnumeration
 from view.show_mode.editor.show_ui_widgets.color_selection_uiwidget import ColorSelectionUIWidget
 from view.show_mode.editor.show_ui_widgets.constant_button_list import ConstantNumberButtonList
 from view.show_mode.editor.show_ui_widgets.cue_control import CueControlUIWidget
-from view.show_mode.editor.show_ui_widgets.pan_tilt_constant import PanTiltConstantControlUIWidget
+from view.show_mode.editor.show_ui_widgets.pan_tilt_constant_show_ui import PanTiltConstantControlUIWidget
 
 
 def filter_to_ui_widget(filter_: Filter, parent_page: "UIPage", configuration: dict[str, str] | None = None,
