+++ src/view/show_mode/editor/show_ui_widgets/pan_tilt_constant_show_ui.py
